safari-534.53-branch/LayoutTests/fast/canvas/canvas-copyPixels-expected.txt

Test if putImageData gives back the same result as getImageData

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS imgdata[4] is 255
PASS imgdata[5] is 0
PASS imgdata[6] is 0
PASS imgdata[4] is 0
PASS imgdata[5] is 128
PASS imgdata[6] is 0
PASS imgdata[4] is 0
PASS imgdata[5] is 0
PASS imgdata[6] is 255
PASS successfullyParsed is true

TEST COMPLETE
